refactor : useInput hook , style : FormDesign.js
